Extract runServer function.
